streamlit-app/requirements.txt
streamlit
llama-index
llama-index-vector-stores-faiss
llama-index-embeddings-huggingface
faiss-cpu